% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{Annot-class}
\alias{Annot-class}
\alias{summary,}
\alias{Annot-method}
\title{Class "Annot"
This class represents the Annotation information}
\description{
Class "Annot"
This class represents the Annotation information
}
\section{Slots}{

\describe{
\item{\code{species}}{the species of the annotation file}

\item{\code{anntype}}{the type of the annotation file}

\item{\code{keytype}}{Gene ID type}

\item{\code{annot}}{Annotation information data.frame}
}}

\author{
Kai Guo
}
\keyword{classes}
