% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barPlot}
\alias{barPlot}
\title{Bar plot for pathway enrichment scores}
\usage{
barPlot(
  object,
  features,
  group_by = NULL,
  color = NULL,
  show.error = TRUE,
  horizontal = FALSE,
  nrow = NULL,
  ncol = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA object or data.frame}

\item{features}{A vector of features (pathways) to plot}

\item{group_by}{Name of metadata column to group cells by}

\item{color}{Colors to use for groups}

\item{show.error}{Show error bars (standard error, default: TRUE)}

\item{horizontal}{Plot bars horizontally (default: FALSE)}

\item{nrow}{Number of rows for faceting}

\item{ncol}{Number of columns for faceting}

\item{basesize}{Base font size}
}
\description{
Creates a bar plot showing mean enrichment scores across groups
}
\examples{
\dontrun{
data(pbmc_small)
hsko <- buildAnnot(species="human", keytype="SYMBOL", anntype="KEGG")
res <- scgsva(pbmc_small, hsko)
barPlot(res, features = c("Wnt.signaling.pathway"), group_by = "groups")
}
}
\author{
Kai Guo
}
