% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{buildAnnot}
\alias{buildAnnot}
\title{make annotation database}
\usage{
buildAnnot(
  species = "human",
  keytype = "SYMBOL",
  anntype = "GO",
  builtin = TRUE,
  OP = NULL
)
}
\arguments{
\item{species}{species for the annotation}

\item{keytype}{the gene Id type to extract}

\item{anntype}{annotation type (GO, KEGG, RO)}

\item{builtin}{use default database (TRUE or FALSE)}

\item{OP}{BP, CC, MF default use all}
}
\description{
Build annotation database
}
\examples{
\dontrun{
annot<-buildAnnot(species = "human", keytype = "SYMBOL",
                  anntype = "KEGG")
}
}
\author{
Kai Guo
}
