% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{buildMSIGDB}
\alias{buildMSIGDB}
\title{Download database from Msigdb and prepare for enrichment analysis}
\usage{
buildMSIGDB(
  species = "human",
  keytype = "SYMBOL",
  anntype = "GO",
  msigdb_data = NULL
)
}
\arguments{
\item{species}{the species for query}

\item{keytype}{the gene ID type}

\item{anntype}{anntotaion type of  gene set (GO,BP,CC,MF,KEGG,REACTOME,
BIOCARTA,HALLMARK)}

\item{msigdb_data}{Optional: pre-loaded msigdbr data frame. Use this if you
cannot connect to Zenodo (e.g., in China). You can download the data manually
and load it with readRDS(), then pass it to this parameter.}
}
\description{
Download database from Msigdb and prepare for enrichment analysis
}
\examples{
\dontrun{
hsamsi<-buildMSIGDB(species = "human", keytype = "SYMBOL", anntype = "GO")

# If you cannot connect to Zenodo, download the file manually:
# 1. Download from: https://zenodo.org/records/15800824/files/msigdb.2025.1.Hs.rds
# 2. Load and use:
# msig_data <- readRDS("msigdb.2025.1.Hs.rds")
# hallmark <- buildMSIGDB(species = "human", keytype = "SYMBOL",
#                         anntype = "HALLMARK", msigdb_data = msig_data)
}
}
\author{
Kai Guo
}
