% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colnames,GSVA-method}
\alias{colnames,GSVA-method}
\alias{colnames,Annot-method}
\title{Get column names of GSVA object}
\usage{
\S4method{colnames}{GSVA}(x)

\S4method{colnames}{Annot}(x)
}
\arguments{
\item{x}{An Annot object}
}
\description{
Get column names of GSVA object

Get column names of Annot object
}
