% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{detail.GSEAResult}
\alias{detail.GSEAResult}
\title{extract the gene information and related pathway}
\usage{
\method{detail}{GSEAResult}(x)
}
\arguments{
\item{x}{GSEAResult object}
}
\description{
extract the gene information and related pathway
}
\examples{
\dontrun{
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
hsako<-as.data.frame(hsako)
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
head(detail(res))
}
}
\author{
Kai Guo
}
