% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{detail.richResult}
\alias{detail.richResult}
\title{extract the gene information and related pathway}
\usage{
\method{detail}{richResult}(x)
}
\arguments{
\item{x}{richResult object}
}
\description{
extract the gene information and related pathway
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  head(detail(res))
}
}
\author{
Kai Guo
}
