% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{.makeGOdata}
\alias{.makeGOdata}
\title{Build GO annotation}
\usage{
.makeGOdata(species = "human", keytype = "ENTREZID", OP = NULL)
}
\arguments{
\item{species}{you can check the support species by using showData()}

\item{keytype}{the gene ID type}

\item{OP}{One of "BP", "MF", and "CC" subontologies, default:"ALL" for
all three.}
}
\description{
Build GO annotation
}
\author{
Kai Guo
}
