% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{.makeKOdata}
\alias{.makeKOdata}
\title{Build KEGG annotation}
\usage{
.makeKOdata(species = "human", keytype = "ENTREZID", builtin = TRUE)
}
\arguments{
\item{species}{you can check the support species by using showData()}

\item{keytype}{the gene ID type}

\item{builtin}{use default database(TRUE or FALSE)}
}
\description{
Build KEGG annotation
}
\author{
Kai Guo
}
