% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{.makeKOMdata}
\alias{.makeKOMdata}
\title{make KEGG module annotation data function}
\usage{
.makeKOMdata(species = "human", keytype = "ENTREZID", builtin = TRUE)
}
\arguments{
\item{species}{you can check the support species by using showData()}

\item{keytype}{the gene ID type}
}
\description{
make KEGG module annotation data function
}
\author{
Kai Guo
}
