% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{enrich}
\alias{enrich}
\title{enrich method}
\usage{
enrich(
  x,
  annot,
  ontology = "",
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  keytype = "",
  filename = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  padj.method = "BH",
  sep = ",",
  ...
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{annot}{ontology type}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{keytype}{keytype for input genes}

\item{filename}{output filename}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}
}
\value{
richResult
}
\description{
enrich
}
\examples{
\dontrun{
library(bioAnno)
fromKEGG(species="ath")
athgo<-buildOwn(dbname="org.ath.eg.db",anntype="GO")
athgo<-as.data.frame(athgo)
athgo<-na.omit(athgo)
gene=sample(unique(athgo$GeneID),1000)
res<-enrich(gene,athgo)
}
}
\author{
Kai Guo
}
