% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{featurePlot}
\alias{featurePlot}
\title{Visualize 'Pathways or Terms' on a dimensional reduction plot}
\usage{
featurePlot(
  object,
  features,
  reduction = "umap",
  color = NULL,
  group_by = NULL,
  label = NULL,
  dims = c(1, 2),
  pt.size = 1,
  pt.shape = 19,
  min.cutoff = NA,
  max.cutoff = NA,
  nrow = NULL,
  ncol = NULL,
  basesize = 12,
  label.size = 4,
  label.color = "black"
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{A vector of features to plot,}

\item{reduction}{Which dimensionality reduction to use. default("umap")}

\item{color}{Colors to use for identity class plotting}

\item{group_by}{Name of one or more metadata columns to group (color) cells by}

\item{label}{Name of one or more metadata columns to label the cells by}

\item{dims}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions}

\item{pt.size}{Size of the points on the plot}

\item{pt.shape}{If NULL, all points are circles (default)}

\item{min.cutoff}{Vector of minimum cutoff values for each feature,
may specify quantile in the form of 'q##' (eg, 'q1', 'q10')}

\item{max.cutoff}{Vector of maximum cutoff values for each feature,
may specify quantile in the form of 'q##' (eg, 'q1', 'q10')}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{basesize}{base font size, given in pts.}

\item{label.size}{Sets the size of the labels}

\item{label.color}{Sets the color of the label text}
}
\description{
Colors single cells on a dimensional reduction plot according to a
'pathways' (i.e. KEGG, GO terms)
}
\examples{
\dontrun{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
res<-scgsva(pbmc_small,hsko)
featurePlot(res,features="Wnt.signaling.pathway")
}
}
\author{
Kai Guo
}
