% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{findPathway}
\alias{findPathway}
\title{Fit linear model for each pathway}
\usage{
findPathway(object, group = NULL, ref = NULL)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{group}{Name of one or more metadata columns to group cells to compare
or a vector if the obj is a data.frame}

\item{ref}{reference group}
}
\description{
Fit linear model for each pathway
}
\examples{
\dontrun{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
sc<-scgsva(pbmc_small,hsko)
res <- findPathway(sc, group = "groups", ref = "g1")
}
}
\author{
Kai Guo
}
