% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{genes,GSVA-method}
\alias{genes,GSVA-method}
\title{gene method}
\usage{
\S4method{genes}{GSVA}(object, features, useTerm = TRUE, with.expr = TRUE)
}
\arguments{
\item{object}{GSVA object}

\item{features}{A vector of features to extract}

\item{useTerm}{use Term or use id (default: TRUE)}

\item{with.expr}{extract the expression value or not (default: TRUE)}
}
\value{
data.frame
}
\description{
gene method
}
\examples{
\dontrun{
data(pbmc_small)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
res<-scgsva(pbmc_small, hsako)
head(genes(res,"Acute myeloid leukemia"))
}
}
\author{
Kai Guo
}
