% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{genes.GSVA}
\alias{genes.GSVA}
\title{extract the gene information and related pathway}
\usage{
\method{genes}{GSVA}(x, features, useTerm = TRUE)
}
\arguments{
\item{x}{richResult object}
}
\description{
extract the gene information and related pathway
}
\examples{
\dontrun{
data(pbmc_small)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
res<-scgsva(pbmc_small, hsako)
head(detail(res))
}
}
\author{
Kai Guo
}
