% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getann}
\alias{getann}
\title{build annotaion for kegg}
\usage{
getann(ontype = "GO")
}
\arguments{
\item{ontype}{GO or KEGG}
}
\description{
build annotaion for kegg
}
\examples{
annot = getann("GO")
}
\author{
Kai Guo
}
