% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getdetail}
\alias{getdetail}
\title{get detail and integrate with the input gene information}
\usage{
getdetail(rese, resd, sep)
}
\arguments{
\item{rese}{richResult or GSEAResult}

\item{resd}{dataframe with input gene as rownames}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
get detail and integrate with the input gene information
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  gened<-data.frame(lfc=rnorm(length(gene)))
  rownames(gened)<-gene
  head(getdetail(res,gened))
}
}
\author{
Kai Guo
}
