% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggnetplot}
\alias{ggnetplot}
\title{ggnetplot method}
\usage{
ggnetplot(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  useTerm = TRUE,
  low = "orange",
  high = "red",
  writeCyt = FALSE,
  cytoscapeFile = NULL,
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = layout.fruchterman.reingold,
  savefig = FALSE,
  filename = NULL,
  width = 7,
  height = 7,
  node.alpha = 0.7,
  repel = TRUE,
  segment.size = 0.2,
  sep = ",",
  ...
)
}
\arguments{
\item{object}{richResult or dataframe}

\item{top}{number of terms to show (default: 50)}

\item{pvalue}{cutoff p value for enrichment result}

\item{padj}{cutoff p adjust value for enrichment result}

\item{usePadj}{use adjust p value for the color or not}

\item{useTerm}{use terms for nodes (default: TRUE)}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{label.color}{label color}

\item{label.size}{label size}

\item{node.shape}{shape of the node}

\item{layout}{layout method}

\item{savefig}{save figures or not}

\item{filename}{output figure name}

\item{width}{width for output figure}

\item{height}{height for output figure}

\item{node.alpha}{alpha-transparency scales}

\item{repel}{use ggrepel text function or not}

\item{segment.size}{segment size for ggrepel text}

\item{sep}{character string used to separate the genes when concatenating}
}
\value{
ggplot2 object
}
\description{
ggnetplot
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggnetplot(res)
}
}
\author{
Kai Guo
}
