% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggnetwork}
\alias{ggnetwork}
\title{ggnetwork method}
\usage{
ggnetwork(
  object,
  gene,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "orange",
  high = "red",
  weightcut = 0.2,
  useTerm = TRUE,
  writeCyt = FALSE,
  cytoscapeFile = NULL,
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = layout.fruchterman.reingold,
  savefig = FALSE,
  visNet = FALSE,
  smooth = TRUE,
  nodeselect = FALSE,
  edit = FALSE,
  savehtml = FALSE,
  filename = NULL,
  width = 7,
  height = 7,
  segment.size = 0.2,
  node.alpha = 0.7,
  sep = ",",
  ...
)
}
\arguments{
\item{object}{richResult,GSEAResult object or dataframe}

\item{gene}{vector contains gene names or dataframe with DEGs information}

\item{top}{number of terms to display}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{weightcut}{cutoff valule for edge}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{label.color}{label color}

\item{label.size}{label size}

\item{layout}{layout for the network (layout.fruchterman.reingold)}

\item{savefig}{save the figure or not}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{segment.size}{size for label segment}

\item{node.alpha}{alpha-transparency scales}

\item{label.font}{label font}
}
\value{
ggplot2 object
}
\description{
ggnetwork
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggnetwork(res)
}
}
