% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{GO_child}
\alias{GO_child}
\title{Get all children terms of node}
\usage{
GO_child(node = "GO:0008150", ontology = "BP")
}
\arguments{
\item{node}{input node of GO}

\item{ontology}{ontology term of BP}
}
\description{
Get all children terms of node
}
\author{
Kai Guo
}
