% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{idconvert}
\alias{idconvert}
\title{Convert id between different type}
\usage{
idconvert(species, keys, fkeytype, tkeytype)
}
\arguments{
\item{species}{species of gene}

\item{keys}{input gene id}

\item{fkeytype}{input gene id type}

\item{tkeytype}{output gene id type}
}
\value{
vector
}
\description{
Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor
annotation package
}
\examples{
\dontrun{
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
hsako<-as.data.frame(hsako)
gene=sample(unique(hsako$GeneID),1000)
id<-idconvert(species="human",fkeytype="SYMBOL",tkeytype="ENTREZID")
}
}
\author{
Kai Guo
}
