% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{lightcolor}
\alias{lightcolor}
\title{light colors for making figures}
\format{
An object of class \code{character} of length 56.
}
\usage{
lightcolor
}
\description{
light colors for making figures
}
\author{
Kai Guo
}
\keyword{datasets}
