% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Lollipop plot for pathway enrichment scores}
\usage{
lollipopPlot(
  object,
  features,
  group_by = NULL,
  color = NULL,
  pt.size = 4,
  horizontal = TRUE,
  order.by = TRUE,
  nrow = NULL,
  ncol = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA object or data.frame}

\item{features}{A vector of features (pathways) to plot}

\item{group_by}{Name of metadata column to group cells by}

\item{color}{Colors to use for groups}

\item{pt.size}{Size of the points}

\item{horizontal}{Plot horizontally (default: TRUE)}

\item{order.by}{Order bars by value (default: TRUE)}

\item{nrow}{Number of rows for faceting}

\item{ncol}{Number of columns for faceting}

\item{basesize}{Base font size}
}
\description{
Creates a lollipop plot showing mean enrichment scores across groups
}
\examples{
\dontrun{
data(pbmc_small)
hsko <- buildAnnot(species="human", keytype="SYMBOL", anntype="KEGG")
res <- scgsva(pbmc_small, hsko)
lollipopPlot(res, features = c("Wnt.signaling.pathway"), group_by = "groups")
}
}
\author{
Kai Guo
}
