% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{msigdbr}
\alias{msigdbr}
\alias{buildMSIGDB}
\title{Download database from Msigdb and prepare for enrichment analysis}
\usage{
buildMSIGDB(
  species = "human",
  keytype = "SYMBOL",
  anntype = "GO",
  category = NULL
)
}
\arguments{
\item{species}{the species for query}

\item{keytype}{the gene ID type}

\item{anntype}{gene set anntype (GO,BP,CC,MF,KEGG,REACTOME,BIOCARTA)}
}
\description{
Download database from Msigdb and prepare for enrichment analysis
}
\examples{
\dontrun{
hsamsi<-buildMSIGDB(species="human",keytype="SYMBOL",anntype="GO")
}
}
\author{
Kai Guo
}
