% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rbind.GSEAResult}
\alias{rbind.GSEAResult}
\title{rbind generic function for GSEAResult object}
\usage{
\method{rbind}{GSEAResult}(...)
}
\description{
rbind generic function for GSEAResult object
}
\author{
Kai Guo
}
