% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rbind.richResult}
\alias{rbind.richResult}
\title{rbind generic function for richResult object}
\usage{
\method{rbind}{richResult}(...)
}
\description{
rbind generic function for richResult object
}
\author{
Kai Guo
}
