% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richCluster}
\alias{richCluster}
\title{kappa cluster analysis}
\usage{
richCluster(
  x,
  gene = NULL,
  useTerm = FALSE,
  cutoff = 0.5,
  overlap = 0.5,
  minSize = 5,
  escore = 3
)
}
\arguments{
\item{x}{richResult object or dataframe}

\item{gene}{(Optional).a vector of gene list}

\item{useTerm}{to use the term or not (TRUE/FALSE)}

\item{cutoff}{kappa score threshold for significant dispersion results}

\item{overlap}{cutoff value of the overlap between two Terms}

\item{minSize}{minimal number of terms in the cluster}

\item{escore}{kappa enrichment score cutoff value (default: 3)}
}
\description{
kappa cluster analysis
}
\examples{
\dontrun{
  hsago<-buildAnnot(species="human",keytype="SYMBOL",anntype = "GO")
  gene=sample(unique(hsago$GeneID),1000)
  res<-richGO(gene,godata = hsago,ontology ="BP")
  resc<-richCluster(res)
}
}
\author{
Kai Guo
}
