% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richGO}
\alias{richGO}
\title{richGO method}
\usage{
richGO(
  x,
  godata,
  ontology = "BP",
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  keytype = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  filename = NULL,
  padj.method = "BH",
  sep = ",",
  ...
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{godata}{GO annotation data}

\item{ontology}{BP,MF or CC}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{keytype}{keytype for input genes}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}
}
\value{
richResult
}
\description{
richGO
}
\examples{
\dontrun{
hsago<-buildAnnot(species="human",keytype="SYMBOL",anntype = "GO")
hsago<-as.data.frame(hsago)
gene=sample(unique(hsago$GeneID),1000)
res<-richGO(gene,godata = hsago,ontology ="BP")
}
}
\author{
Kai Guo
}
