% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richGSEA}
\alias{richGSEA}
\title{richGSEA method}
\usage{
richGSEA(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  minSize = 15,
  ontology = "",
  maxSize = 500,
  padj.method = "BH",
  organism = NULL,
  table = TRUE,
  sep = ",",
  ...
)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{object}{annotation file for all genes}

\item{pvalue}{pvalue cutoff value}

\item{padj}{adjust p value cut off method}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{padj.method}{p value adjust method}

\item{table}{leadingEdge as vector}

\item{sep}{character string used to separate the genes when concatenating}
}
\value{
GSEAResult
}
\description{
richGSEA
}
\examples{
\dontrun{
set.seed(123)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
}
}
\author{
Kai Guo
}
