% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richKEGG}
\alias{richKEGG}
\title{richKEGG method}
\usage{
richKEGG(
  x,
  kodata,
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  ontology = "KEGG",
  keytype = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  filename = NULL,
  padj.method = "BH",
  builtin = TRUE,
  sep = ",",
  ...
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{kodata}{GO annotation data}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{ontology}{KEGG}

\item{keytype}{keytype for input genes}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}

\item{bulitin}{use KEGG bulit in KEGG annotation or not(set FALSE if you want use newest KEGG data)}
}
\value{
richResult
}
\description{
richKEGG
}
\examples{
\dontrun{
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
gene=sample(unique(hsako$GeneID),1000)
res<-richKEGG(gene,kodata = hsako)
}
}
\author{
Kai Guo
}
