% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rownames,GSVA-method}
\alias{rownames,GSVA-method}
\alias{rownames,Annot-method}
\title{Get row names of GSVA object}
\usage{
\S4method{rownames}{GSVA}(x)

\S4method{rownames}{Annot}(x)
}
\arguments{
\item{x}{An Annot object}
}
\description{
Get row names of GSVA object

Get row names of Annot object
}
