% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{Small SingleCellExperiment object}
\format{
A \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\source{
\url{https://scrnaseq-public-datasets.s3.amazonaws.com/scater-objects/segerstolpe.rds}
}
\usage{
sce_small
}
\description{
Small SingleCellExperiment object
}
\examples{
data("sce_small")
}
\concept{data}
\keyword{datasets}
