% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set.cut.off}
\alias{set.cut.off}
\title{modified from Seurat github
transform data with the min and max cutoff}
\usage{
set.cut.off(gsva, min.cutoff = NA, max.cutoff = NA)
}
\arguments{
\item{gsva}{GSVA result}

\item{min.cutoff}{Vector of minimum cutoff values for each feature,
may specify quantile in the form of 'q##' (eg, 'q1', 'q10')}

\item{max.cutoff}{Vector of maximum cutoff values for each feature,
may specify quantile in the form of 'q##' (eg, 'q1', 'q10')}
}
\value{
data.frame
}
\description{
modified from Seurat github
transform data with the min and max cutoff
}
