% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{showData}
\alias{showData}
\title{show avaliable data based on bioconductor annotation package}
\usage{
showData()
}
\description{
show avaliable data based on bioconductor annotation package
}
\author{
Kai Guo
}
