% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{sigPathway}
\alias{sigPathway}
\title{Significance testing between groups.}
\usage{
sigPathway(
  object,
  group = NULL,
  test.use = "wilcox",
  ref = NULL,
  method = "BH"
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{group}{Name of one or more metadata columns to group cells to compare
or a vector if the obj is a data.frame}

\item{test.use}{Statistical test to use: "wilcox" (default), "t.test", or "anova"}

\item{ref}{reference group}

\item{method}{correction method, a character string}
}
\description{
Significance testing between groups.
}
\examples{
\dontrun{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
sc<-scgsva(pbmc_small,hsko)
res <- sigPathway(sc, group = "groups")
}
}
\author{
Kai Guo
}
