% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_data_matrix}
\alias{split_data_matrix}
\title{Split data matrix into smaller sub-matrices ('chunks')}
\usage{
split_data_matrix(matrix, chunk.size = 1000)
}
\arguments{
\item{matrix}{Input data matrix}

\item{chunk.size}{How many cells to include in each sub-matrix}
}
\value{
A list of sub-matrices, each with size (n_features x chunk_size)
}
\description{
Split data matrix into smaller sub-matrices ('chunks')
}
