% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset.GSVA}
\alias{subset.GSVA}
\title{subset the GSVA object}
\usage{
\method{subset}{GSVA}(x, features = NULL, cells = NULL, ...)
}
\arguments{
\item{x}{a GSVA object}

\item{features}{character vector of pathway/feature names to keep in gsva slot}

\item{cells}{character vector of cell barcodes to keep}

\item{...}{additional arguments passed to subset.Seurat (excluding features)}
}
\description{
subset the GSVA object
}
