% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{buildAnnot}
\alias{buildAnnot}
\title{make annotation database}
\usage{
buildAnnot(
  species = "human",
  keytype = "SYMBOL",
  anntype = "GO",
  builtin = TRUE,
  OP = NULL
)
}
\arguments{
\item{species}{species for the annotation}

\item{keytype}{key type export}

\item{anntype}{annotation type}

\item{OP}{BP,CC,MF default use all}

\item{bulitin}{use default database(TRUE or FALSE)}
}
\description{
build annotation database
}
\examples{
\dontrun{
annot<-buildAnnot(species="human",keytype="ENTREZID",anntype="GO",bulitin=TRUE)
}
}
\author{
Kai Guo
}
