% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.do_ttest}
\alias{.do_ttest}
\title{do t.test}
\usage{
.do_ttest(x, group, ref = NULL, method = "BH", ...)
}
\arguments{
\item{x}{data.frame with sample id as the column name, genes or otu as rownames}

\item{group}{group factor used for comparison}

\item{ref}{reference group}

\item{method}{correction method, a character string}

\item{...}{parameters to t_test}
}
\description{
do t.test
}
\examples{
{
data("mtcars")
do_ttest(mtcars,group="vs")
do_ttest(mtcars,group="cyl",ref="4")
}
}
\author{
Kai Guo
}
