% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get.quantile}
\alias{.get.quantile}
\title{get the quantile of data}
\usage{
.get.quantile(cutoff, data)
}
\description{
get the quantile of data
}
\examples{
set.seed(42)
get.quantile('q10', sample(1:100, 10))

}
