% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{.makeKOdata}
\alias{.makeKOdata}
\title{make KEGG annotation data function}
\usage{
.makeKOdata(species = "human", keytype = "ENTREZID", builtin = TRUE)
}
\arguments{
\item{species}{you can check the support species by using showData()}

\item{keytype}{the gene ID type}
}
\description{
make KEGG annotation data function
}
\author{
Kai Guo
}
