% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnot.R
\name{.makeROdata}
\alias{.makeROdata}
\title{make Reactome annotation data function}
\usage{
.makeROdata(species = "human", keytype = "SYMBOL")
}
\arguments{
\item{species}{you can check the supported species by using showAvailableRO}

\item{keytype}{key type export}
}
\description{
make Reactome annotation data function
}
\author{
Kai Guo
}
