% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dotPlot}
\alias{dotPlot}
\title{DotPlot for the specific pathway}
\usage{
dotPlot(
  object,
  features,
  group_by = NULL,
  split.by = NULL,
  color = NULL,
  pt.size = 3,
  pt.shape = 19,
  nrow = NULL,
  ncol = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{A vector of features to plot}

\item{group_by}{Name of one or more metadata columns to group (color) cells by}

\item{split.by}{Factor to split the groups by}

\item{color}{Colors to use for identity class plotting}

\item{pt.size}{Size of the points on the plot}

\item{pt.shape}{If NULL, all points are circles (default)}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{basesize}{base font size, given in pt}
}
\description{
Intuitive way of visualizing how pathway changes across
different identity classes (clusters).
}
\examples{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
res<-scgsva(pbmc_small,hsko)
dotPlot(res,features="Wnt.signaling.pathway")
}
\author{
Kai Guo
}
