% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{findPathway}
\alias{findPathway}
\title{Fit linear model for each pathway}
\usage{
findPathway(object, group = NULL, ref = NULL, pvalue = 0.05, padj = NULL)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{group}{Name of one or more metadata columns to group cells to compare
or a vector if the obj is a data.frame}

\item{ref}{reference group}

\item{pvalue}{pvalue cut off}

\item{padj}{padj cut off}
}
\description{
Fit linear model for each pathway
}
\author{
Kai Guo
}
