% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{genes}
\alias{genes}
\title{gene method}
\usage{
genes(object, features, useTerm = TRUE, with.expr = TRUE)
}
\arguments{
\item{object}{GSVA object}

\item{features}{A vector of features to extract}

\item{useTerm}{use Term or use id (default: TRUE)}

\item{with.expr}{extract the expression value or not (default: TRUE)}
}
\value{
data.frame
}
\description{
genes
}
\author{
Kai Guo
}
