% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{idconvert}
\alias{idconvert}
\title{Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package}
\usage{
idconvert(species, keys, fkeytype, tkeytype)
}
\arguments{
\item{species:}{you can check the support species by using showData()}

\item{fkeytype:}{the gene type you want to convert}

\item{tkeytype:}{the gene type you want to get}
}
\description{
Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  gene=sample(unique(hsako$GeneID),1000)
  id<-idconvert(species="human",fkeytype="SYMBOL",tkeytype="ENTREZID")
}
}
\author{
Kai Guo

Kai Guo
}
