% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ridgePlot}
\alias{ridgePlot}
\title{Generate a ridge plot to examine enrichment distributions}
\usage{
ridgePlot(
  object,
  features,
  group_by = NULL,
  color = NULL,
  split.by = NULL,
  rug = TRUE,
  nrow = NULL,
  ncol = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{A vector of features to plot}

\item{group_by}{Name of one or more metadata columns to group (color) cells by
or a vector to show the group}

\item{color}{Colors to use for identity class plotting}

\item{split.by}{Factor to split the groups by}

\item{rug}{Adds a rug representation or not}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{basesize}{base font size, given in pt}
}
\description{
This function allows to the user to generate the distribution of
enrichment across groups with a ridge plot.
}
\examples{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
res<-scgsva(pbmc_small,hsko)
ridgePlot(res,features="Wnt.signaling.pathway")
}
\author{
Kai Guo
}
