% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{scgsva}
\alias{scgsva}
\title{GSVA function for single cell data or data.frame with expression value}
\usage{
scgsva(
  obj,
  annot = NULL,
  method = "ssgsea",
  kcdf = "Poisson",
  abs.ranking = FALSE,
  min.sz = 1,
  max.sz = Inf,
  mx.diff = TRUE,
  ssgsea.norm = TRUE,
  useTerm = TRUE,
  cores = 4,
  verbose = TRUE
)
}
\arguments{
\item{obj}{The count matrix, Seurat, or SingleCellExperiment object.}

\item{annot}{annotation object}

\item{method}{to employ in the estimation of gene-set enrichment scores per
sample. By default this is set to gsva}

\item{kcdf}{Character string denoting the kernel to use during the
non-parametric estimation of the cumulative distribution function of
expression levels across samples when method="gsva".
By default, kcdf="Poisson"}

\item{abs.ranking}{Flag used only when mx.diff=TRUE.}

\item{min.sz}{Minimum size of the resulting gene sets}

\item{max.sz}{Maximum size of the resulting gene sets.}

\item{mx.diff}{Offers two approaches to calculate the enrichment
statistic (ES) from the KS random walk statistic.}

\item{ssgsea.norm}{Logical, set to TRUE (default) with method="ssgsea"
runs the SSGSEA method}

\item{useTerm}{use Term or use id (default: TRUE)}

\item{cores}{The number of cores to use for parallelization.}

\item{verbose}{Gives information about each calculation step. Default: FALSE.}
}
\description{
GSVA function for single cell data or data.frame with expression value
}
\author{
Kai Guo
}
