% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{scgsva}
\alias{scgsva}
\title{GSVA function for single cell data or data.frame with expression value}
\usage{
scgsva(
  obj,
  annot = NULL,
  cores = 4,
  method = "ssgsea",
  kcdf = "Poisson",
  abs.ranking = FALSE,
  min.sz = 1,
  max.sz = Inf,
  mx.diff = TRUE,
  ssgsea.norm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{The count matrix, Seurat, or SingleCellExperiment object.}

\item{cores}{The number of cores to use for parallelization.}

\item{annotation}{annotation object}
}
\description{
GSVA function for single cell data or data.frame with expression value
}
\author{
Kai Guo
}
