% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spatialFeaturePlot}
\alias{spatialFeaturePlot}
\title{Visualize 'Pathways or Terms' on a Spatial  plot}
\usage{
spatialFeaturePlot(
  object,
  features,
  images = NULL,
  color = NULL,
  label = NULL,
  pt.size = 1,
  pt.shape = 19,
  min.cutoff = NA,
  max.cutoff = NA,
  nrow = NULL,
  ncol = NULL,
  basesize = 12,
  label.size = 4,
  label.color = "black"
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{Name of the feature to visualize.}

\item{images}{Name of the images to use in the plot(s)}

\item{color}{Colors to use for identity class plotting}

\item{label}{Name of one or more metadata columns to label the cells by}

\item{pt.size}{Size of the points on the plot}

\item{pt.shape}{If NULL, all points are circles (default)}

\item{min.cutoff, }{max.cutoff Vector of minimum and maximum cutoff values for
each feature, may specify quantile in the form of 'q##' where
is the quantile (eg, 'q1', 'q10')}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{basesize}{base font size, given in pts.}

\item{label.size}{Sets the size of the labels}

\item{label.color}{Sets the color of the label text}
}
\description{
Colors single cells on a spatial plot according to a
'pathways' (i.e. KEGG, GO terms)
}
\examples{
set.seed(123)
library(scGSVA)
library(Seurat)
library(SeuratData)
InstallData("stxBrain")
brain <- LoadData("stxBrain", type = "anterior1")
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
res<-scgsva(brain,hsko)
spatialFeaturePlot(res,features="Wnt.signaling.pathway")
}
\author{
Kai Guo
}
