% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset.GSVA}
\alias{subset.GSVA}
\title{subset the GSVA object}
\usage{
\method{subset}{GSVA}(x, ...)
}
\arguments{
\item{x}{a GSVA object}

\item{...}{arguments used in subset.Seurat}
}
\description{
subset the GSVA object
}
