% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vlnPlot}
\alias{vlnPlot}
\title{VlnPlot for the pathways}
\usage{
vlnPlot(
  object,
  features,
  group_by = NULL,
  split.by = NULL,
  color = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{A vector of features to plot}

\item{group_by}{Name of one or more metadata columns to group (color) cells by}

\item{split.by}{Factor to split the groups by}

\item{color}{Colors to use for identity class plotting}

\item{basesize}{base font size, given in pts.}
}
\description{
Draws a violin plot of single cell data (KEGG, GO)
}
\author{
Kai Guo
}
