% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vlnPlot}
\alias{vlnPlot}
\title{VlnPlot for the pathways}
\usage{
vlnPlot(
  object,
  features,
  group_by = NULL,
  color = NULL,
  split.by = NULL,
  split.plot = FALSE,
  pt.size = 0,
  pt.shape = 19,
  nrow = NULL,
  ncol = NULL,
  basesize = 12
)
}
\arguments{
\item{object}{A GSVA objectect or data.frame}

\item{features}{A vector of features to plot}

\item{group_by}{Name of one or more metadata columns to group (color) cells by}

\item{color}{Colors to use for identity class plotting}

\item{split.by}{Factor to split the groups by}

\item{split.plot}{plot each group of the split violin plots by multiple or
single violin shapes.}

\item{pt.size}{Size of the points on the plot}

\item{pt.shape}{If NULL, all points are circles (default)}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{basesize}{base font size, given in pts.}
}
\description{
Draws a violin plot of single cell data (KEGG, GO)
}
\examples{
set.seed(123)
library(scGSVA)
data(pbmc_small)
hsko<-buildAnnot(species="human",keytype="SYMBOL",anntype="KEGG")
res<-scgsva(pbmc_small,hsko)
vlnPlot(res,features="Wnt.signaling.pathway")
}
\author{
Kai Guo
}
